Fix check if product exists in other combinations
Theme used SUM(combination.quantity) to determine if product can be
ordered for at least one combination.

This is not correct, for example when we have 3 combinations with
quantities (1, -3, 0). The sum would be -2, yet there is one combination
that can be ordered.

The fix is to calculate sum of positive quantities instead.